+++ tutorial/video/src/knapsack_objective.cpp
@@ -1,8 +1,8 @@
 #include "knapsack_objective.hpp"
 
-KSObjective::KSObjective( const std::vector<ghost::Variable>& variables, std::shared_ptr<ghost::AuxiliaryData> data )
+KSObjective::KSObjective( const std::vector<ghost::Variable>& variables, const std::vector<int>&& values )
 	: Maximize( variables, "Max profit" ),
-	  _values( std::dynamic_pointer_cast<KSCoefficients>(data)->values )
+	  _values( std::move( values ) )
 { }
 
 double KSObjective::required_cost( const std::vector<ghost::Variable*>& variables ) const