+++ ChangeLog.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file, since GHOST 2.0.0.
 
+## [2.1.0] - 2021-10-04
+This is a minor version change due to some interface modifications.
+
+- Permutation problem declaration is now done in the `ModelBuilder` constructor.
+- Remove `Objective::expert_postprocess_satisfaction` and change the name and the interface of `Objective::expert_postprocess` (previously `Objective::expert_postprocess_optimization`)
+
 ## [2.0.1] - 2021-09-30
 - Fix a bad implementation of the copy-and-swap idiom in ghost::Options
 - Fix some typos in the Solver screen outputs when the macro GHOST_BENCH is declared